requirements.txt
streamlit==1.35.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.34
sentence-transformers==2.7.0
chromadb==0.4.24
numpy==1.26.4
tiktoken==0.7.0
duckduckgo-search==5.3.1b1
PyMuPDF==1.24.1
python-dotenv==1.0.1streamlit==1.35.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.34
sentence-transformers==2.7.0
numpy==1.26.4
tiktoken==0.7.0
duckduckgo_search
PyMuPDF==1.24.1
python-dotenv==1.0.1

# --- NEW LIBRARIES FOR FAISS ---
faiss-cpu 
# --- REMOVE CHROMA ---
# chromadb==0.4.24 (delete this line)